Added a button for uploading picture to parse if a picture is user is done selecting a photo
